HSEARCH-3118 Move ContextualErrorHandler to the engine
... so that we can use it for Lucene too.
