+++ engine/src/main/java/org/hibernate/search/engine/common/spi/DefaultContextualErrorHandler.java
@@ -4,11 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.elasticsearch.orchestration.impl;
-
-import org.hibernate.search.backend.elasticsearch.work.impl.ElasticsearchWork;
-import org.hibernate.search.engine.common.spi.ErrorContextBuilder;
-import org.hibernate.search.engine.common.spi.ErrorHandler;
+package org.hibernate.search.engine.common.spi;
 
 /**
  * Registers the errors during an execution ultimately passes them to an error handler.
@@ -25,26 +21,23 @@ class DefaultContextualErrorHandler implements ContextualErrorHandler {
 
 	private Throwable errorThatOccurred;
 
-	public DefaultContextualErrorHandler(ErrorHandler errorHandler) {
-		super();
+	DefaultContextualErrorHandler(ErrorHandler errorHandler) {
 		this.errorHandler = errorHandler;
 	}
 
 	@Override
-	public void markAsFailed(ElasticsearchWork<?> work, Throwable throwable) {
-		Object info = work.getInfo();
-		if ( info != null ) {
-			getErrorContextBuilder().operationAtFault( info );
-			getErrorContextBuilder().addWorkThatFailed( info );
+	public void markAsFailed(Object workInfo, Throwable throwable) {
+		if ( workInfo != null ) {
+			getErrorContextBuilder().operationAtFault( workInfo );
+			getErrorContextBuilder().addWorkThatFailed( workInfo );
 		}
 		addThrowable( throwable );
 	}
 
 	@Override
-	public void markAsSkipped(ElasticsearchWork<?> work) {
-		Object info = work.getInfo();
-		if ( info != null ) {
-			getErrorContextBuilder().addWorkThatFailed( info );
+	public void markAsSkipped(Object workInfo) {
+		if ( workInfo != null ) {
+			getErrorContextBuilder().addWorkThatFailed( workInfo );
 		}
 	}
